+++ src/Components/CaughtPokemon/CaughtPokemonContainer.js
@@ -1,0 +1,14 @@
+import React from "react";
+import CaughtPokemon from "./CaughtPokemon";
+
+
+class CaughtPokemonContainer extends React.Component {
+    render() {
+        return (
+        <CaughtPokemon />
+        )
+    }
+}
+
+
+export default CaughtPokemonContainer;
