--- src/Components/DELETECard/DELETECard.js
@@ -1,16 +1,0 @@
-// import React from "react";
-
-// const Card = (props) => {
-
-//     return (
-//         <div className="card" >
-//             <img src={`pokemons/${props.id}.png`} className="card-img-top" alt="image" />
-//             <div className="card-body">
-//                 <h5 className="card-title">{props.name.toUpperCase()}</h5>
-//                 <a href="#" className="btn btn-primary">CAUGHT</a>
-//             </div>
-//         </div>
-//     )
-// }
-
-// export default Card;
