+++ src/Components/PokemonList/PokemonList.js
@@ -1,31 +1,22 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
+import Pagination from './../Pagination/Pagination';
 
 let PokemonList = (props) => {
-
-    let pagesCount = Math.ceil(props.totalPokemonsCount / props.pageSize);
-
-    let pages = [];
-    for (let i=1; i<=pagesCount; i++) {
-        pages.push(i);
-    }
-
-    
     return (
         <div className="container-xl">
-            {/* pagination */}
-            <div>
-                {pages.map ( p => {
-                    return <button type="button" className={props.currentPage === p ? "btn btn-success" : "btn btn-outline-success"} 
-                    onClick = { () => {props.onPageChanged(p)} } > {p} </button>
-                })}
-            </div>
-            {/* pokemon cards */}
+            <Pagination 
+                totalPokemonsCount={props.totalPokemonsCount}
+                pageSize={props.pageSize}
+                currentPage={props.currentPage}
+                onPageChanged={props.onPageChanged}
+                />
+                
             <div className="row">
                 {props.pokemons.map(pokemonItem => {
                     return (
                         <div className="col-md-3" key={pokemonItem.id}>
-                            <div className="card" >
+                            <div className="card mt-3" >
                                 <NavLink to={'/profile/' + pokemonItem.id}>
                                     <img src={`/pokemons/${pokemonItem.id}.png`} className="card-img-top" alt="image" />
                                 </NavLink>
