--- src/redux/DELETEstore_data.js
@@ -1,49 +1,0 @@
-// // import Data from "../db.json";
-// // let store = Data.pokemons;
-
-// let store = {
-//     pokemons: [
-//       {
-//         name: "bulbasaur",
-//         id: 1
-//       },
-//       {
-//         name: "ivysaur",
-//         id: 2
-//       },
-//       {
-//         name: "venusaur",
-//         id: 3
-//       },
-//       {
-//         name: "charmander",
-//         id: 4
-//       },
-//       {
-//         name: "charmeleon",
-//         id: 5
-//       },
-//       {
-//         name: "charizard",
-//         id: 6
-//       },
-//       {
-//         name: "squirtle",
-//         id: 7
-//       },
-//       {
-//         name: "wartortle",
-//         id: 8
-//       },
-//       {
-//         name: "blastoise",
-//         id: 9
-//       },
-//       {
-//         name: "caterpie",
-//         id: 10
-//       }
-//     ]
-// }
-
-// export default store;
